Add Memory Access Verifier
This commit introduces a memory access verifier that checks for
miscompilations and sandbox escapes, such as [1].

The tool is very much a prototype and 'research quality', i.e. there are
many improvements that can be made.

Nevertheless, if [1] is reintroduced, fuzzing will find the bug after
some time.

See docs/mem-access-verifier.md for an overview how this tool works.

This commit is a squashed commit of ~230 commits made over the last two
months.

[1]: https://github.com/bytecodealliance/wasmtime/security/advisories/GHSA-ff4p-7xrq-q5r8